+++ gp_testing.py
@@ -22,7 +22,7 @@ from gpytorch.priors.lkj_prior import LKJCovariancePrior
 from gpytorch.priors.smoothed_box_prior import SmoothedBoxPrior
 import math
 from gpytorch.constraints import GreaterThan
-torch.set_default_dtype(torch.float64)
+torch.set_default_dtype(torch.float32)
 from pydantic import BaseModel, Field
 from typing import List, Literal
 from datetime import date
@@ -179,16 +179,16 @@ scaler_y2.fit(y2)
 X_scaled = scaler_X.transform(X_train)
 y_scaled = scaler_y.transform(y_train)
 y2_scaled = scaler_y2.transform(y_train2)
-train_x = torch.tensor(X_scaled, dtype=torch.float64)
-train_y = torch.tensor(y_scaled, dtype=torch.float64).flatten()
-train_y2 = torch.tensor(y2_scaled, dtype=torch.float64).flatten()
+train_x = torch.tensor(X_scaled, dtype=torch.float32)
+train_y = torch.tensor(y_scaled, dtype=torch.float32).flatten()
+train_y2 = torch.tensor(y2_scaled, dtype=torch.float32).flatten()
 
 X_test_scaled = scaler_X.transform(X_test)
-test_x = torch.tensor(X_test_scaled, dtype=torch.float64)
+test_x = torch.tensor(X_test_scaled, dtype=torch.float32)
 y_test_scaled = scaler_y.transform(y_test)
 y_test_scaled2 = scaler_y2.transform(y_test2)
-test_y = torch.tensor(y_test_scaled, dtype=torch.float64).flatten()
-test_y2 = torch.tensor(y_test_scaled2, dtype=torch.float64).flatten()
+test_y = torch.tensor(y_test_scaled, dtype=torch.float32).flatten()
+test_y2 = torch.tensor(y_test_scaled2, dtype=torch.float32).flatten()
 
 ##################Single Task GP######################
 
@@ -277,9 +277,9 @@ for i in range(training_iter):
         break  # Stop training early if patience is exceeded
 
 X_full = scaler_X.transform(X)
-x_tensor = torch.tensor(X_full, dtype=torch.float64)
+x_tensor = torch.tensor(X_full, dtype=torch.float32)
 y_full = scaler_y.transform(y)
-y_tensor = torch.tensor(y_full, dtype=torch.float64).flatten()
+y_tensor = torch.tensor(y_full, dtype=torch.float32).flatten()
 
 # Get into evaluation (predictive posterior) mode
 model.eval()
@@ -472,13 +472,13 @@ f, (y1_ax, y2_ax) = plt.subplots(1, 2, figsize=(12, 5))
 
 
 X_full = scaler_X.transform(X)
-x_tensor = torch.tensor(X_full, dtype=torch.float64)
+x_tensor = torch.tensor(X_full, dtype=torch.float32)
 test_i_task1 = torch.full((x_tensor.shape[0],1), dtype=torch.long, fill_value=0)
 test_i_task2 = torch.full((x_tensor.shape[0],1), dtype=torch.long, fill_value=1)
 y_full = scaler_y.transform(y)
 y_full2 = scaler_y2.transform(y2)
-y_tensor = torch.tensor(y_full, dtype=torch.float64).flatten()
-y_tensor2 = torch.tensor(y_full2, dtype=torch.float64).flatten()
+y_tensor = torch.tensor(y_full, dtype=torch.float32).flatten()
+y_tensor2 = torch.tensor(y_full2, dtype=torch.float32).flatten()
 
 
 # Make predictions - one task at a time
@@ -691,11 +691,11 @@ likelihood.eval()
 f, (y1_ax, y2_ax) = plt.subplots(1, 2, figsize=(12, 5))
 
 X_full = scaler_X.transform(X)
-x_tensor = torch.tensor(X_full, dtype=torch.float64)
+x_tensor = torch.tensor(X_full, dtype=torch.float32)
 y_full = scaler_y.transform(y)
 y_full2 = scaler_y2.transform(y2)
-y_tensor = torch.tensor(y_full, dtype=torch.float64).flatten()
-y_tensor2 = torch.tensor(y_full2, dtype=torch.float64).flatten()
+y_tensor = torch.tensor(y_full, dtype=torch.float32).flatten()
+y_tensor2 = torch.tensor(y_full2, dtype=torch.float32).flatten()
 y_tensor_comb = torch.stack([y_tensor, y_tensor2], -1)
 
 # Make predictions
